requirements.txt
beautifulsoup4
dataclasses==0.6
dataclasses-json==0.6.7
fastapi==0.113.0
langchain-google-genai
loguru
python-dotenv==1.0.1
requests==2.32.3
ulid
uvicorn==0.30.6
uvloop==0.19.0
websockets==12.0
